Fix an issue where aliases were run out of order
This was happening because Queue_PushFront was being invoked for each command, meaning they would end up in the queue backwards.

To fix, we just iterate through the line backwards. This gives us exactly what we need, even recursively, because when we invoke the unpacked commands, they too are unpacked in reverse, and so on.